Extract publishing and animalsniffer plugins.
