+++ exporters/jaeger/build.gradle.kts
@@ -1,8 +1,8 @@
 plugins {
     id("otel.protobuf-conventions")
-    `maven-publish`
+    id("otel.publish-conventions")
 
-    id("ru.vyarus.animalsniffer")
+    id("otel.animalsniffer-conventions")
 }
 
 description = "OpenTelemetry - Jaeger Exporter"
