+++ exporters/otlp/trace/build.gradle.kts
@@ -1,10 +1,10 @@
 plugins {
     id("otel.java-conventions")
-    `maven-publish`
+    id("otel.publish-conventions")
 
     id("otel.jmh-conventions")
     id("org.unbroken-dome.test-sets")
-    id("ru.vyarus.animalsniffer")
+    id("otel.animalsniffer-conventions")
 }
 
 description = "OpenTelemetry Protocol Trace Exporter"
