+++ sdk-extensions/aws/build.gradle.kts
@@ -1,6 +1,6 @@
 plugins {
     id("otel.java-conventions")
-    `maven-publish`
+    id("otel.publish-conventions")
 
     id("otel.jmh-conventions")
 }
