+++ sdk-extensions/jfr-events/build.gradle.kts
@@ -1,6 +1,6 @@
 plugins {
     id("otel.java-conventions")
-    `maven-publish`
+    id("otel.publish-conventions")
 }
 
 description = "OpenTelemetry SDK Extension JFR"
