+++ sdk/metrics/build.gradle.kts
@@ -1,12 +1,12 @@
 plugins {
     id("otel.java-conventions")
-    id("maven-publish")
+    id("otel.publish-conventions")
 
     id("otel.jmh-conventions")
 
     // TODO(anuraaga): Enable animalsniffer by the time we are getting ready to release a stable
     // version. Long/DoubleAdder are not part of Android API 21 which is our current target.
-    // id("ru.vyarus.animalsniffer")
+    // id("otel.animalsniffer-conventions")
 }
 
 description = "OpenTelemetry SDK Metrics"
