+++ settings.gradle.kts
@@ -9,7 +9,6 @@ pluginManagement {
         id("org.checkerframework") version "0.5.20"
         id("org.jetbrains.kotlin.jvm") version "1.5.10"
         id("org.unbroken-dome.test-sets") version "4.0.0"
-        id("ru.vyarus.animalsniffer") version "1.5.3"
     }
 }
